Moved PasswordReset model class out of model.py
