+++ iot_api/user_api/models/__init__.py
@@ -21,3 +21,4 @@ from iot_api.user_api.models.AppKey import AppKey
 from iot_api.user_api.models.ChangeEmailRequests import ChangeEmailRequests
 from iot_api.user_api.models.GlobalData import GlobalData
 from iot_api.user_api.models.LoginAttempts import LoginAttempts
+from iot_api.user_api.models.PasswordReset import PasswordReset
