+++ iot_api/user_api/resources/endpoints.py
@@ -28,7 +28,7 @@ from iot_api.user_api.enums import RoleTypes
 from iot_api.user_api.events.data_collector_events import emit_data_collector_event
 from iot_api.user_api.events.policy_events import emit_policy_event
 from iot_api.user_api.model import User, Organization, Device, RevokedTokenModel, AccountActivation, StatsCounters, \
-    PasswordReset, UserRole, UserToUserRole, Alert, AlertType, Packet, \
+    UserRole, UserToUserRole, Alert, AlertType, Packet, \
     SendMailAttempts, \
     get_user_collector_ids
 from iot_api.user_api.models.ChangeEmailRequests import ChangeEmailRequests
@@ -42,6 +42,7 @@ from iot_api.user_api.models.NotificationAlertSettings import NotificationAlertS
 from iot_api.user_api.models.NotificationData import NotificationData
 from iot_api.user_api.models.NotificationDataCollectorSettings import NotificationDataCollectorSettings
 from iot_api.user_api.models.NotificationPreferences import NotificationPreferences
+from iot_api.user_api.models.PasswordReset import PasswordReset
 from iot_api.user_api.models.Policy import Policy
 from iot_api.user_api.model import Quarantine
 from iot_api.user_api.singletonURL import singletonURL
